ADDED DATA CLEANING STEPS, CORRELATION ANALYSIS, AND PCA
Now, the WISE dataset is filtered for CH, for years, where
data is available, data is interpolated where possible, the
time window is narrowed down to 1995 - 2020. Then I perform
a pairwise correlation analysis and a PCA to figure out how
many PCs are needed to explain most of the variance. Then I
check if I can win back some interpretability by figuring out
if an original index is highly aligned with the main PC.