ADDED TIME SERIES-RELATED ANALYSIS
Previously, the correlation analysis was performed on the raw
index values. However, this may have lead to spurious correlations
because of temporal trends. To fix this, I now perform the correlation
analysis based on percentage differences between the values of each
index/time series. This analysis has been run and the related results
have been stored.