INDICATOR-INDEPENDENT RESTRUCTURING/CLEANING OF DATA TABLES
The data tables downloaded from the web are not cleanly formatted
for data analysis as there is text in the tables. Now the code
handles this and restructures the tables such that the output are
tables with proper indices and column headers. The explanatory
remarks are extracted and compiled into a string separately.