REFACTORING & DEBUGGING DataFileLoader CLASS
The DataFileLoader was a bit buggy as some bugs were introduced when
copy-pasting code from original drafts to final code file. These
bugs were now fixed.

Additionally, the processed data files are now stored as json files
instead of csv flat files. This change was introduced in order to keep
meta data such as damid, observable name, remarks, descriptions etc.
in the file. All this info was lost when storing to flatfiles (CSV).

Design choice argument:
All the MONET2030 datafiles are quite small
(<10 KB per file) and there are only just over 100 files. This is not
expected to change anytime soon. As a result, we're not dealing with
a lot of data. JSON is therefore a simple, lightweight solution to
handle the usecase of keeping the data together with the metainfo.