FIXING BUGS & ADDED FORCE-SCRAPING FEATURE
- Fixed a typo
- Fixed logic bug: The test of how many data files are expected to
be in the folder was wrong. Now the code correctly checks if the
right number of data files is in the folder (as of July 1, 2025 there
should be 125 files) and reads them from the disk if they are there.
Otherwise it scrapes the files from the web.
- Force-scraping feature: I now also introduced a feature that allows
the user to force the code to scrape the files even if the correct
number of files is already on disk.