REFACTORING - FILE STURCTURE IMPROVEMENT
The meta data files are now better structured. They are now fully
relational, i.e. there are primary keys in all data sets (set to
be in the first column) which can be related to foreign keys
in the other data sets. Further, I added meta information files
logging the actual data file download and processing. They contain
information about when a specific file was downloaded or processed,
respectively, they contain a file hash for quickly checking if the
content of a file has changed etc.