ADDED 3RD PROCESSING STEP
This 3rd processing step is outside of the ETL/ELT process used
in this project. The reason is that the ETL/ELT is defined on a
per-index or per-observable level. The 3rd processing step, in
contrast, compiles all the information from all indices, observables,
and metrics. This is why this processing step is defined in
monet_scraper.py and not in monet_etl.py.

In this 3rd processing step, we compile two big tables: one containing
the metric values (columns) for all years recorded (rows) and a
second one with the equivalent structure but containing 95%
confidence intervals.

For most of the subsequent analysis, the compiled metrics table
will be the dataset to work with.