REFACTOR - TIME SERIES ANALYSIS
I already detrended the time series previously in order to compute
the correlations between them based on the residuals/deviations
from the trend. However, there was a problem because I interpolated
the time series (in order to get data points over a common time grid
for all metrics) linearly and detrended them using first differences.
As a result, the residuals were piecewise constant. As many metrics
were measured at similar points in time, the intervals over which
the first differences were constant were often very similar leading
to high but spurious correlations.

I'm trying to fix this now with a different approach:
(1) I use Gaussian Processes to interpolate the data --> interpolation
	is no longer linear --> differences are no longer constant
(2) I detrend the time series by actually fitting a trend instead
	of simply taking the first differences.The trend window smoothing
	length is optimized for each metric individually such that the
	p-value of the augmented Dickey-Fuller (ADF) test on the resulting
	residuals becomes minimal (i.e. the residual becomes maximally
	stationary).

Notice that there are several strong (and not fully qualified) assumptions
going into this strategy:
a) I use a Matérn kernel in the GPs with a length scale of 3.3 because this
   leads to visually reasonable uncertainty envelopes (although without further
   subject matter expertise this is hard to quantify). Then again, the
   uncertainty envelopes are not really relevant for the rest of the analysis,
   i.e. a wrong length scale has no significant impact.
b) Using the p-vale of the ADF test as the objective function to optimize
   during detrending assumes that optimal stationarity translates directly
   to optimal trend, which may not be true. This may actually lead to over-fitted.
   This is in fact an aspect one should not forget an come back to if issues
   occur further down the road.
c) When ultimately counting how many metrics could be made stationary I use a
   significance level alpha=10%, which is quite high. However, given the somewhat
   hand-wavey arguments above this is likely not going to have a strong detrimental
   impact (10% is still a commonly used value for alpha).