START OF REFACTORING: MOVING TRANSFORMATIONS TO SEPARATE SUBMODULE
Until now, the ELT process was basically entirely covered in the
monet_elt.py and monet_scraper.py submodules. I have decided to
move the transformations completely to a separate submodule which
will group the transformations together more neatly and will make
it more natural to keep track/version each and every step along
the data transformation pipeline.

The original motivation for this was the realization that data
transformation does not stop just because the data files reach
a format that is analysis-friendly. Transformations continue in
the form of data cleaning, data imputation etc. For full control,
data should be stored at every step along the transformation chain.
This is made possible with the newly introduced processor submodule.