MINOR REFACTOR
When raw MONET data files are read into memory now, this is done
in a for loop now and no longer in a list comprehension. While this
may have a negative impact on performance, it allows to drop the
"Unnamed: 0" column (former index) from all worksheets in one go.
Notice that we cannot just use the "usecols" option in pd.read_excel
in order to avoid explicitly dropping the "Unnamed: 0" column as
we're dealing with raw files here for which we don't know the column
headers (which had to pass to "usecols").