NOTEBOOK CONTAINING EXPLORATORY DATA ANALYSIS (EDA)
This notebook explores the MONET 2030 data both as a whole as well
as split by capital. It visualizes temporal trends in each time
series, measures and visualizes summary statistics (box plots, etc.)
and filters out the 3 top and the 3 worst performing metrics (again
both overall as well as in a per-capital approach). It counts the
number of official MONET 2030 key indicators among the both groups.
Last but not least it ranks the key indicators by their performance.