DATA PROCESSING PIPELINE HAS NOW CHECKPOINTING & IS RESUMABLE
At this point each data processing step is represented as
a class in the monet_processor.py submodule. All these
steps are then consolidated into the TransformationPipeline
class. This class has a method "run", which is the only method
the user needs to interact with. The "run" method automatically
performs checks which of all the data transformation steps is the
last step for which data is already available on disk. This data
is loaded into memory and used to run through the remaining
processing steps.

Through the use of the kwarg 'force' in the initializer of the
TransformationPipeline class the user can also force the recomputation
of any of the transformation steps even if the corresponding data
should already be available on disk.