BUGFIXES RELATED TO CHECKPOINTING & RESUMING
Previously, when a stage n was forced to be recomputed, this did
not cascade down to subsequent stages. This is now fixed. Notice
that this makes sense: If, e.g. stage 2 is forced to be recomputed
it makes sense to recompute also the following stages as otherwise
the recomputed results from stage 2 would be ignored in the rest
of the pipeline.

In the course of fixing the bug above, I also realized that there
was a bug in the private _read methods of the individual Stage
classes: in them the output attribute was not properly defined. This
used to lead to the situation where self.output and self.input
were empty as soon as one stage was read from disk (i.e. as soon
as a _read method was called) instead of computed (calling _transform).
This is also fixed now.