REFACTORED "RUN" LOGIC
The previous "run" method is now called "resume" and "run" has
be re-designed not as a recursive function running through a stack
of data transformation stages but rather as looping over them.

This change was made as I realized that when I only resume from the
latest data available on disk I miss all the results from earlier
stages in memory, even though I might need access to them later.

Therefore, the code of "run" now simply runs through all data
transformation stages starting from the 1st stage and loads the from
disk if it is available already or computes it, if it is not. This way,
all the results will be loaded into memory from all stages.