UPDATED TIME SERIES ANALYSIS CODE
The time series detrending is now only performed if
that analysis step is either enforced by the user
or if the corresponding data is not already available
on disk. If the later is the case, however, and the
user does not force re-execution of the detrending,
the results are merely loaded into memory from disk.