7TH DATA TRANSFORMATION STEP: Z-SCALING
Standard scaling was added as the 7th stage in the
data transformation pipeline. Currently, it is only
applied to the time series residuals due to the linear
design of the transformation pipeline.

YET TO DO: Implement functionality to apply the scaling
ALSO to the imputed (but non-detrended) time series.
Notice that this implies a deviation from the linear
transformation pipeline design and would be the start
to a more graph-based transformation setup.