MAJOR REFACTOR: SCRAPER
The scraper is now more object-oriented and the different scraping
steps are now encapsulated into a class "Scraper". The functionality
is exposed to the user through the "scrape" method. As a result,
the user can now run all scraping steps by simply instantiating
the "Scraper" class and run it via "Scraper.scrape()".