UPDATED ID2NAME MAP IN DATA TRANSFORMATION STAGE 2
Now, the capitals information is also stored in this id2name
map.

Further more, the TransformationPipeline's run function now has
a new optional argument "stage_id". If this parameter is set
to an integer between 1 and len(stages)+1, then the corresponding
stage within the TransforamtionPipeline is run (but no subsequent
stages are run).