RENAMED SCRAPER CLASS AND SCRAPER METHOD
Now the class is called MonetLoader and the user interface method
is called "load" in alignment with the WISE code base. This makes
more sense as the method only checks if scraping is even required.
If the data is already available on disk, the method really only
reads the data into memory (it loads it).