REFACTORING METATABLE SCRAPING
The scraper module now also loads the key indicators and merges
this information AND the capital mappings into the indicator meta
table. This information is then propagated into the observables
meta table.

YET TO DO: Key indicator scraper

The key indicator information is currently only read from disk, i.e.
the file has to exists. This also holds for the capital mapping,
although that one will be very tricky to scrape from the internet
as this would have to be semantics-based using definitions of both
the capitals and the individual metrics (--> would require LLM).