ADDED CORRELATION ANALYSIS FOR KEY METRICS ONLY
Recall that the goal of the correlation analysis is to reduce
the number of metrics. The set of key indicators as defined by
MONET/Swiss federal office of statistics is already a reduction
of the list of metrics. Therefore, one could also take this
as a starting point for the correlation analysis. That's exactly
what's been added now.