UPDATED PATH HANDLING
The file paths can now be configured in advance by setting
a iso3 country code in the analysis notebook. This country
code defines for which country the data analysis shall be
conducted. However, it also creates new subfolders (if they
don't exist already) where the country-specific results will
be stored.